+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v5.2.0]
 ### Added
 - Support for CRAM alignment format when creating IGV config files.
 - Support for an optional "analysis_groups" column in the sample sheet providing grouping information for the input samples.